#96 use double to obtain the current material properties

+++ bouncer/src/repo/manipulator/modelconvertor/import/repo_model_import_assimp.cpp
@@ -243,7 +243,7 @@ repo::core::model::MaterialNode* AssimpModelImport::createMaterialRepoNode(
 		repo_material_t repo_material;
 
 		aiColor3D tempColor;
-		float tempFloat;
+		double tempFloat;
 
 		//--------------------------------------------------------------------------
 		// Ambient
